requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.1
imgaug>=0.4.0

# MLOps & Experiment Tracking
mlflow>=2.7.0
dvc>=3.0.0
dvc-gdrive>=2.20.0
wandb>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# API & Deployment
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Monitoring
prometheus-client>=0.17.0
# Dataset download utility
kagglehub